refactor: RequestDTO 적용 - Party
