+++ src/main/java/kr/flab/ottsharing/controller/PartyController.java
@@ -7,14 +7,15 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import kr.flab.ottsharing.dto.request.PartyCreateDto;
+import kr.flab.ottsharing.dto.request.PartyIdDto;
+import kr.flab.ottsharing.dto.request.PartyUpdateDto;
 import kr.flab.ottsharing.dto.response.MyParty;
 import kr.flab.ottsharing.dto.response.PartyCreateResult;
 import kr.flab.ottsharing.dto.response.PartyDeleteResult;
 import kr.flab.ottsharing.dto.response.PartyJoinResult;
-import kr.flab.ottsharing.dto.response.UpdatePartyInfo;
 import kr.flab.ottsharing.service.PartyService;
 import lombok.RequiredArgsConstructor;
 
@@ -24,8 +25,8 @@ public class PartyController {
     private final PartyService partyServ;
 
     @PostMapping("/party/create")
-    public PartyCreateResult create(@CookieValue(name = "userId") String userId, @RequestParam String ottId, @RequestParam String ottPassword) {
-        return partyServ.create(userId, ottId, ottPassword);
+    public PartyCreateResult create(@CookieValue(name = "userId") String userId, @RequestBody PartyCreateDto createDto) {
+        return partyServ.create(userId, createDto.getOttId(), createDto.getOttPassword());
     }
 
     @GetMapping("/myParty")
@@ -39,17 +40,17 @@ public class PartyController {
     }
 
     @DeleteMapping("/party/deleteParty")
-    public PartyDeleteResult deleteParty(@CookieValue(name = "userId") String userId, @RequestParam Integer partyId) {
-        return partyServ.deleteParty(userId, partyId);
+    public PartyDeleteResult deleteParty(@CookieValue(name = "userId") String userId, @RequestBody PartyIdDto partyIdDto) {
+        return partyServ.deleteParty(userId, partyIdDto.getPartyId());
     }
 
     @DeleteMapping("/party/getOutParty")
-    public String getOutParty(@CookieValue(name = "userId") String userId, @RequestParam Integer partyId) {
-        return partyServ.getOutParty(userId, partyId);
+    public String getOutParty(@CookieValue(name = "userId") String userId, @RequestBody PartyIdDto partyIdDto) {
+        return partyServ.getOutParty(userId, partyIdDto.getPartyId());
     }
 
     @PostMapping("/party/updatePartyInfo")
-    public String updatePartyInfo(@CookieValue(name = "userId") String userId, @Valid @RequestBody UpdatePartyInfo info) {
+    public String updatePartyInfo(@CookieValue(name = "userId") String userId, @Valid @RequestBody PartyUpdateDto info) {
         return partyServ.updatePartyInfo(userId, info);
     }
 }
