+++ src/main/java/kr/flab/ottsharing/dto/request/PartyCreateDto.java
@@ -1,0 +1,9 @@
+package kr.flab.ottsharing.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class PartyCreateDto {
+    private String ottId;
+    private String ottPassword;
+}
