+++ src/main/java/kr/flab/ottsharing/dto/request/PartyIdDto.java
@@ -1,0 +1,8 @@
+package kr.flab.ottsharing.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class PartyIdDto {
+    private Integer partyId;
+}
